requirements.txt
openai
autogen
langgraph
playwright
pytest
python-dotenv
